refactor sdk usage
