+++ packages/sdk/scripts/move-artifacts.sh
@@ -2,14 +2,9 @@
 rm -rf ./artifacts
 mkdir ./artifacts
 
-# define the array of files ( should be excluded)
-EXCLUDED_FILES=(
-  "./lib/contracts/out/ICErc20.sol/ICErc20.json"
-)
 
-for file in $(find ./lib/contracts/out -name '*.json' | grep -iv "${EXCLUDED_FILES[@]}" | grep -iv test); do
-    cp $file ./artifacts;
-done
-
+cp -R './lib/contracts/out/' ./artifacts # --exclude ./lib/contracts/out/*Test*
+rm -rf ./artifacts/*Test*.sol
+rm -rf ./artifacts/*.t.sol
 
 cp ./lib/contracts/out/test.sol/*.json ./artifacts;
