+++ packages/sdk/src/IonicSdk/index.ts
@@ -15,25 +15,25 @@ import {
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { BigNumber, Contract, Signer, utils } from "ethers";
 
-import CTokenFirstExtensionABI from "../../abis/CTokenFirstExtension";
-import EIP20InterfaceABI from "../../abis/EIP20Interface";
-import FeeDistributorABI from "../../abis/FeeDistributor";
-import IonicERC4626ABI from "../../abis/IonicERC4626";
-import IonicFlywheelLensRouterABI from "../../abis/IonicFlywheelLensRouter";
-import IonicLiquidatorABI from "../../abis/IonicLiquidator";
-import PoolDirectoryABI from "../../abis/PoolDirectory";
-import PoolLensABI from "../../abis/PoolLens";
-import PoolLensSecondaryABI from "../../abis/PoolLensSecondary";
-import UnitrollerABI from "../../abis/Unitroller";
+import { abi as CTokenFirstExtensionABI } from "../../artifacts/CTokenFirstExtension.sol/CTokenFirstExtension.json";
+import { abi as EIP20InterfaceABI } from "../../artifacts/EIP20Interface.sol/EIP20Interface.json";
+import { abi as FeeDistributorABI } from "../../artifacts/FeeDistributor.sol/FeeDistributor.json";
+import { abi as IonicERC4626ABI } from "../../artifacts/IonicERC4626.sol/IonicERC4626.json";
+import { abi as IonicFlywheelLensRouterABI } from "../../artifacts/IonicFlywheelLensRouter.sol/IonicFlywheelLensRouter.json";
+import { abi as IonicLiquidatorABI } from "../../artifacts/IonicLiquidator.sol/IonicLiquidator.json";
+import { abi as PoolDirectoryABI } from "../../artifacts/PoolDirectory.sol/PoolDirectory.json";
+import { abi as PoolLensABI } from "../../artifacts/PoolLens.sol/PoolLens.json";
+import { abi as PoolLensSecondaryABI } from "../../artifacts/PoolLensSecondary.sol/PoolLensSecondary.json";
+import { abi as UnitrollerABI } from "../../artifacts/Unitroller.sol/Unitroller.json";
 import { CTokenFirstExtension } from "../../typechain/CTokenFirstExtension";
 import { EIP20Interface } from "../../typechain/EIP20Interface";
-import { FeeDistributor } from "../../typechain/FeeDistributor";
+import { FeeDistributor } from "../../typechain/FeeDistributor.sol/FeeDistributor";
 import { IonicERC4626 as IonicERC4626 } from "../../typechain/IonicERC4626";
-import { IonicFlywheelLensRouter as IonicFlywheelLensRouter } from "../../typechain/IonicFlywheelLensRouter";
+import { IonicFlywheelLensRouter as IonicFlywheelLensRouter } from "../../typechain/IonicFlywheelLensRouter.sol/IonicFlywheelLensRouter";
 import { IonicLiquidator } from "../../typechain/IonicLiquidator";
 import { PoolDirectory } from "../../typechain/PoolDirectory";
 import { PoolLens } from "../../typechain/PoolLens";
-import { PoolLensSecondary } from "../../typechain/PoolLensSecondary";
+import { PoolLensSecondary } from "../../typechain/PoolLensSecondary.sol/PoolLensSecondary";
 import { Unitroller } from "../../typechain/Unitroller";
 import { withAsset } from "../modules/Asset";
 import { withConvertMantissa } from "../modules/ConvertMantissa";
