+++ packages/sdk/src/modules/CreateContracts.ts
@@ -1,36 +1,36 @@
 import { Contract, ContractInterface } from "ethers";
 
 import { IonicBaseConstructor } from "..";
-import AuthoritiesRegistryABI from "../../abis/AuthoritiesRegistry";
-import CompoundMarketERC4626ABI from "../../abis/CompoundMarketERC4626";
-import FlywheelStaticRewardsABI from "../../abis/FlywheelStaticRewards";
-import ICErc20ABI from "../../abis/ICErc20";
-import ICErc20PluginRewardsABI from "../../abis/ICErc20PluginRewards";
-import ILeveredPositionFactoryABI from "../../abis/ILeveredPositionFactory";
-import ILiquidatorsRegistryABI from "../../abis/ILiquidatorsRegistry";
-import IonicComptrollerABI from "../../abis/IonicComptroller";
-import IonicFlywheelABI from "../../abis/IonicFlywheel";
-import IonicFlywheelLensRouterABI from "../../abis/IonicFlywheelLensRouter";
-import JumpRateModelABI from "../../abis/JumpRateModel";
-import LeveredPositionABI from "../../abis/LeveredPosition";
-import LeveredPositionsLensABI from "../../abis/LeveredPositionsLens";
-import MasterPriceOracleABI from "../../abis/MasterPriceOracle";
-import OptimizedAPRVaultFirstExtensionABI from "../../abis/OptimizedAPRVaultFirstExtension";
-import OptimizedAPRVaultSecondExtensionABI from "../../abis/OptimizedAPRVaultSecondExtension";
-import OptimizedVaultsRegistryABI from "../../abis/OptimizedVaultsRegistry";
-import PoolLensSecondaryABI from "../../abis/PoolLensSecondary";
-import PoolRolesAuthorityABI from "../../abis/PoolRolesAuthority";
-import UnitrollerABI from "../../abis/Unitroller";
+import { abi as AuthoritiesRegistryABI } from "../../artifacts/AuthoritiesRegistry.sol/AuthoritiesRegistry.json";
+import { abi as CompoundMarketERC4626ABI } from "../../artifacts/CompoundMarketERC4626.sol/CompoundMarketERC4626.json";
+import { abi as IonicComptrollerABI } from "../../artifacts/ComptrollerInterface.sol/IonicComptroller.json";
+import { abi as ICErc20PluginRewardsABI } from "../../artifacts/CTokenInterfaces.sol/CErc20PluginRewardsInterface.json";
+import { abi as FlywheelStaticRewardsABI } from "../../artifacts/FlywheelStaticRewards.sol/FlywheelStaticRewards.json";
+import { abi as ICErc20ABI } from "../../artifacts/ICErc20.sol/ICErc20.json";
+import { abi as ILeveredPositionFactoryABI } from "../../artifacts/ILeveredPositionFactory.sol/ILeveredPositionFactory.json";
+import { abi as ILiquidatorsRegistryABI } from "../../artifacts/ILiquidatorsRegistry.sol/ILiquidatorsRegistry.json";
+import { abi as IonicFlywheelABI } from "../../artifacts/IonicFlywheel.sol/IonicFlywheel.json";
+import { abi as IonicFlywheelLensRouterABI } from "../../artifacts/IonicFlywheelLensRouter.sol/IonicFlywheelLensRouter.json";
+import { abi as JumpRateModelABI } from "../../artifacts/JumpRateModel.sol/JumpRateModel.json";
+import { abi as LeveredPositionABI } from "../../artifacts/LeveredPosition.sol/LeveredPosition.json";
+import { abi as LeveredPositionsLensABI } from "../../artifacts/LeveredPositionsLens.sol/LeveredPositionsLens.json";
+import { abi as MasterPriceOracleABI } from "../../artifacts/MasterPriceOracle.sol/MasterPriceOracle.json";
+import { abi as OptimizedAPRVaultFirstExtensionABI } from "../../artifacts/OptimizedAPRVaultFirstExtension.sol/OptimizedAPRVaultFirstExtension.json";
+import { abi as OptimizedAPRVaultSecondExtensionABI } from "../../artifacts/OptimizedAPRVaultSecondExtension.sol/OptimizedAPRVaultSecondExtension.json";
+import { abi as OptimizedVaultsRegistryABI } from "../../artifacts/OptimizedVaultsRegistry.sol/OptimizedVaultsRegistry.json";
+import { abi as PoolLensSecondaryABI } from "../../artifacts/PoolLensSecondary.sol/PoolLensSecondary.json";
+import { abi as PoolRolesAuthorityABI } from "../../artifacts/PoolRolesAuthority.sol/PoolRolesAuthority.json";
+import { abi as UnitrollerABI } from "../../artifacts/Unitroller.sol/Unitroller.json";
 import { AuthoritiesRegistry } from "../../typechain/AuthoritiesRegistry";
 import { CompoundMarketERC4626 } from "../../typechain/CompoundMarketERC4626";
+import { IonicComptroller } from "../../typechain/ComptrollerInterface.sol/IonicComptroller";
+import { ICErc20 } from "../../typechain/CTokenInterfaces.sol/ICErc20";
+import { ICErc20PluginRewards } from "../../typechain/CTokenInterfaces.sol/ICErc20PluginRewards";
 import { FlywheelStaticRewards } from "../../typechain/FlywheelStaticRewards";
-import { ICErc20 } from "../../typechain/ICErc20";
-import { ICErc20PluginRewards } from "../../typechain/ICErc20PluginRewards";
-import { ILeveredPositionFactory } from "../../typechain/ILeveredPositionFactory";
-import { ILiquidatorsRegistry } from "../../typechain/ILiquidatorsRegistry";
-import { IonicComptroller } from "../../typechain/IonicComptroller";
+import { ILeveredPositionFactory } from "../../typechain/ILeveredPositionFactory.sol/ILeveredPositionFactory";
+import { ILiquidatorsRegistry } from "../../typechain/ILiquidatorsRegistry.sol/ILiquidatorsRegistry";
 import { IonicFlywheel } from "../../typechain/IonicFlywheel";
-import { IonicFlywheelLensRouter } from "../../typechain/IonicFlywheelLensRouter";
+import { IonicFlywheelLensRouter } from "../../typechain/IonicFlywheelLensRouter.sol/IonicFlywheelLensRouter";
 import { JumpRateModel } from "../../typechain/JumpRateModel";
 import { LeveredPosition } from "../../typechain/LeveredPosition";
 import { LeveredPositionsLens } from "../../typechain/LeveredPositionsLens";
@@ -38,7 +38,7 @@ import { MasterPriceOracle } from "../../typechain/MasterPriceOracle";
 import { OptimizedAPRVaultFirstExtension } from "../../typechain/OptimizedAPRVaultFirstExtension";
 import { OptimizedAPRVaultSecondExtension } from "../../typechain/OptimizedAPRVaultSecondExtension";
 import { OptimizedVaultsRegistry } from "../../typechain/OptimizedVaultsRegistry";
-import { PoolLensSecondary } from "../../typechain/PoolLensSecondary";
+import { PoolLensSecondary } from "../../typechain/PoolLensSecondary.sol/PoolLensSecondary";
 import { PoolRolesAuthority } from "../../typechain/PoolRolesAuthority";
 import { Unitroller } from "../../typechain/Unitroller";
 import { SignerOrProvider } from "../IonicSdk";
