+++ packages/sdk/src/modules/Vaults.ts
@@ -1,7 +1,7 @@
 import { FlywheelRewardsInfoForVault, FundOperationMode, SupportedChains, VaultData } from "@ionicprotocol/types";
 import { BigNumber, constants, ContractTransaction, utils } from "ethers";
 
-import EIP20InterfaceABI from "../../abis/EIP20Interface";
+import { abi as EIP20InterfaceABI } from "../../artifacts/EIP20Interface.sol/EIP20Interface.json";
 import { getContract } from "../IonicSdk/utils";
 
 import { CreateContractsModule } from "./CreateContracts";
