+++ packages/sdk/tasks/admin/revenue.ts
@@ -5,7 +5,7 @@ import { BigNumber, Contract, providers } from "ethers";
 import { task, types } from "hardhat/config";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { default as ERC20Abi } from "../../abis/EIP20Interface";
+import { abi as ERC20Abi } from "../../artifacts/EIP20Interface.sol/EIP20Interface.json";
 import { IonicERC4626 } from "../../typechain/IonicERC4626";
 import { PoolDirectory } from "../../typechain/PoolDirectory";
 
@@ -25,7 +25,7 @@ async function cgPrice(cgId: string) {
 }
 
 async function createComptroller(
-  pool: PoolDirectory.IonicPoolStructOutput,
+  pool: PoolDirectory.PoolStructOutput,
   deployer: SignerWithAddress
 ): Promise<ComptrollerWithExtension | null> {
   const ionicSdkModule = await import("../ionicSdk");
