+++ packages/sdk/tasks/vaults/deploy.ts
@@ -1,7 +1,7 @@
 import { constants } from "ethers";
 import { task, types } from "hardhat/config";
 
-import { CErc20 } from "../../typechain/CErc20";
+import { CErc20Delegate } from "../../typechain/CErc20Delegate";
 import { CompoundMarketERC4626 } from "../../typechain/CompoundMarketERC4626";
 import { OptimizedAPRVaultFirstExtension } from "../../typechain/OptimizedAPRVaultFirstExtension";
 import { OptimizedAPRVaultSecondExtension } from "../../typechain/OptimizedAPRVaultSecondExtension";
@@ -14,7 +14,7 @@ export default task("deploy-optimized:all")
     let asset;
     const markets = marketsAddresses.split(",");
     for (let i = 0; i < markets.length; i++) {
-      const cErc20 = (await ethers.getContractAt("CErc20", markets[i])) as CErc20;
+      const cErc20 = (await ethers.getContractAt("CErc20", markets[i])) as CErc20Delegate;
       const marketUnderlying = await cErc20.callStatic.underlying();
       if (!asset) asset = marketUnderlying;
       if (asset != marketUnderlying) throw new Error(`The vault adapters should be for the same underlying`);
