+++ packages/sdk/tasks/vaults/rewards.ts
@@ -5,9 +5,9 @@ import { CErc20RewardsDelegate } from "../../typechain/CErc20RewardsDelegate";
 import { Comptroller } from "../../typechain/Comptroller";
 import { ComptrollerFirstExtension } from "../../typechain/ComptrollerFirstExtension";
 import { IERC20MetadataUpgradeable as IERC20 } from "../../typechain/IERC20MetadataUpgradeable";
-import { IERC20Mintable } from "../../typechain/IERC20Mintable";
 import { IonicFlywheel } from "../../typechain/IonicFlywheel";
 import { MasterPriceOracle } from "../../typechain/MasterPriceOracle";
+import { IERC20Mintable } from "../../typechain/MockLpTokenStaker.sol/IERC20Mintable";
 import { OptimizedAPRVaultFirstExtension } from "../../typechain/OptimizedAPRVaultFirstExtension";
 import { OptimizedVaultsRegistry } from "../../typechain/OptimizedVaultsRegistry";
 import { SimplePriceOracle } from "../../typechain/SimplePriceOracle";
